+++ CSE/declaration_public.vb
@@ -52,7 +52,7 @@ Module declaration_public
 
     ' Constances for the array declaration
     Public JumpPoint As List(Of Integer)                        ' Point at that a jump in the time-resolved data is detected
-    Public OptPar() As Boolean = ({True, True, True})           ' Array to identify if optional parameters are given
+    Public OptPar() As Boolean = ({True, True, True, True})     ' Array to identify if optional parameters are given
     Public AT As Boolean = False                                ' Calculation of an automatic transmission (with n_card)
     Public MT_AMT As Boolean = False                            ' Calculation of an manual transmission (always when n_eng available)