+++ CHANGES.md
@@ -5,7 +5,11 @@ TUG improvements:
 
   * Bugfix heading control
   * Update user manual
-  * Input "Report bug" button (at the moment this does nothing)
+  * Add "Report bug" option
+  * Add JIRA Quick Start Guide
+  * Delete valid_t_tire from output files and job-file
+  * Change t_tire from required in input data to optional (like p_tire)
+  * Delete rho_air_ref from criteria and GUI
   
 #### 2015-07-01: v2.0.2-beta6 ####
 TUG improvements: